Adding GCC attributes to increase warnings
git-svn-id: https://svn.parrot.org/parrot/trunk@18727 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/exceptions.c
@@ -198,7 +198,7 @@ Parrot_pop_mark(Interp *interp, INTVAL mark)
             = stack_entry(interp, interp->dynamic_env, 0);
         if (!e)
             real_exception(interp, NULL, 1,
-                           "Mark %d not found.", mark);
+                           "Mark %ld not found.", (long)mark);
         (void)stack_pop(interp, &interp->dynamic_env,
                         NULL, e->entry_type);
         if (e->entry_type == STACK_ENTRY_MARK) {
